Enum pallet_elections::Error

pub enum Error<T: Trait> {
    NotVoter,
    InactiveTarget,
    CannotReapPresenting,
    ReapGrace,
    NotProxy,
    InvalidReporterIndex,
    InvalidTargetIndex,
    InvalidVoteIndex,
    CannotRetractPresenting,
    RetractNonVoter,
    InvalidRetractionIndex,
    DuplicatedCandidate,
    InvalidCandidateSlot,
    InsufficientCandidateFunds,
    InsufficientPresenterFunds,
    ZeroDeposit,
    UnworthyCandidate,
    LeaderboardMustExist,
    NotPresentationPeriod,
    InvalidCandidate,
    DuplicatedPresentation,
    IncorrectTotal,
    InvalidVoterIndex,
    InsufficientVoterFunds,
    InsufficientLockedValue,
    TooManyVotes,
    ZeroCandidates,
    ApprovalPresentation,
    // some variants omitted
}

Error for the elections module.

Variants

NotVoter

Reporter must be a voter.

InactiveTarget

Target for inactivity cleanup must be active.

CannotReapPresenting

Cannot reap during presentation period.

ReapGrace

Cannot reap during grace period.

NotProxy

Not a proxy.

InvalidReporterIndex

Invalid reporter index.

InvalidTargetIndex

Invalid target index.

InvalidVoteIndex

Invalid vote index.

CannotRetractPresenting

Cannot retract when presenting.

RetractNonVoter

Cannot retract non-voter.

InvalidRetractionIndex

Invalid retraction index.

DuplicatedCandidate

Duplicate candidate submission.

InvalidCandidateSlot

Invalid candidate slot.

InsufficientCandidateFunds

Candidate has not enough funds.

InsufficientPresenterFunds

Presenter must have sufficient slashable funds.

ZeroDeposit

Stake deposited to present winner and be added to leaderboard should be non-zero.

UnworthyCandidate

Candidate not worthy of leaderboard.

LeaderboardMustExist

Leaderboard must exist while present phase active.

NotPresentationPeriod

Cannot present outside of presentation period.

InvalidCandidate

Presented candidate must be current.

DuplicatedPresentation

Duplicated presentation.

IncorrectTotal

Incorrect total.

InvalidVoterIndex

Invalid voter index.

InsufficientVoterFunds

New voter must have sufficient funds to pay the bond.

InsufficientLockedValue

Locked value must be more than limit.

TooManyVotes

Amount of candidate votes cannot exceed amount of candidates.

ZeroCandidates

Amount of candidates to receive approval votes should be non-zero.

ApprovalPresentation

No approval changes during presentation period.