Module pallet_elections::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

GenesisConfig
Can be used to configure the genesis state of this pallet.

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageApprovalsOf

_GeneratedPrefixForStorageCandidateCount

_GeneratedPrefixForStorageCandidates

_GeneratedPrefixForStorageDesiredSeats

_GeneratedPrefixForStorageLeaderboard

_GeneratedPrefixForStorageMembers

_GeneratedPrefixForStorageNextFinalize

_GeneratedPrefixForStorageNextVoterSet

_GeneratedPrefixForStoragePresentationDuration

_GeneratedPrefixForStorageRegisterInfoOf

_GeneratedPrefixForStorageTermDuration

_GeneratedPrefixForStorageVoteCount

_GeneratedPrefixForStorageVoterCount

_GeneratedPrefixForStorageVoterInfoOf

_GeneratedPrefixForStorageVoters

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

Traits

Config
Configuration trait of this pallet.

Type Definitions

ApprovalsOf
TWOX-NOTE: SAFE as AccountId is a crypto hash and SetIndex is not attacker-controlled.

CandidateCount
Current number of active candidates

Candidates
The present candidate list.

DesiredSeats
Number of accounts that should constitute the collective.

Leaderboard
Get the leaderboard if we’re in the presentation phase. The first element is the weight of each entry; It may be the direct summed approval stakes, or a weighted version of it. Sorted from low to high.

Members
The current membership. When there’s a vote going on, this should still be used for executive matters. The block number (second element in the tuple) is the block that their position is active until (calculated by the sum of the block number when the member was elected and their term duration).

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

NextFinalize
The accounts holding the seats that will become free on the next tally.

NextVoterSet
the next free set to store a voter in. This will keep growing.

PresentationDuration
How long to give each top candidate to present themselves after the vote ends.

RegisterInfoOf
The vote index and list slot that the candidate who was registered or None if they are not currently registered.

TermDuration
How long each position is active for.

VoteCount
The total number of vote rounds that have happened or are in progress.

VoterCount
Current number of Voters.

VoterInfoOf
Basic information about a voter.

Voters
The present voter list (chunked and capped at VOTER_SET_SIZE).