Struct pallet_elections::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn set_approvals(
origin: OriginFor<T>,
votes: Vec<bool>,
index: VoteIndex,
hint: u32,
value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResult

Set candidate approvals. Approval slots stay valid as long as candidates in those slots are registered.

Locks value from the balance of origin indefinitely. Only [retract_voter] or [reap_inactive_voter] can unlock the balance.

hint argument is interpreted differently based on:

if origin is setting approvals for the first time: The index will be checked for being a valid hole in the voter list.

if the hint is correctly pointing to a hole, no fee is deducted from origin.

Otherwise, the call will succeed but the index is ignored and simply a push to the last chunk with free space happens. If the new push causes a new chunk to be created, a fee indicated by [VotingFee] is deducted.

if origin is already a voter: the index must be valid and point to the correct position of the origin in the current voters list.

Note that any trailing false votes in votes is ignored; In approval voting, not voting for a candidate and voting false, are equal.

O(1).

Two extra DB entries, one DB change.

Argument votes is limited in length to number of candidates.

pub fn reap_inactive_voter(
origin: OriginFor<T>,
reporter_index: u32,
who: <T::Lookup as StaticLookup>::Source,
who_index: u32,
assumed_vote_index: VoteIndex
) -> DispatchResult

Remove a voter. For it not to be a bond-consuming no-op, all approved candidate indices must now be either unregistered or registered to a candidate that registered the slot after the voter gave their last approval set.

Both indices must be provided as explained in [voter_at] function.

May be called by anyone. Returns the voter deposit to signed.

O(1).

Two fewer DB entries, one DB change.

pub fn retract_voter(origin: OriginFor<T>, index: u32) -> DispatchResult

Remove a voter. All votes are cancelled and the voter deposit is returned.

The index must be provided as explained in [voter_at] function.

Also removes the lock on the balance of the voter. See [do_set_approvals()].

O(1).

Two fewer DB entries, one DB change.

pub fn submit_candidacy(origin: OriginFor<T>, slot: u32) -> DispatchResult

Submit oneself for candidacy.

Account must have enough transferrable funds in it to pay the bond.

NOTE: if origin has already assigned approvals via [set_approvals], it will NOT have any usable funds to pass candidacy bond and must first retract. Note that setting approvals will lock the entire balance of the voter until retraction or being reported.

Independent of input.

Three DB changes.

pub fn present_winner(
origin: OriginFor<T>,
candidate: <T::Lookup as StaticLookup>::Source,
total: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
index: VoteIndex
) -> DispatchResult

Claim that candidate is one of the top carry_count + desired_seats candidates. Only works iff the presentation period is active. candidate should have at least collected some non-zero total votes and origin must have enough funds to pay for a potential slash.

O(voters) compute.

One DB change.

pub fn set_desired_seats(origin: OriginFor<T>, count: u32) -> DispatchResult

Set the desired member count; if lower than the current count, then seats will not be up election when they expire. If more, then a new vote will be started if one is not already in progress.

pub fn remove_member(
origin: OriginFor<T>,
who: <T::Lookup as StaticLookup>::Source
) -> DispatchResult

Remove a particular member from the set. This is effective immediately.

Note: A tally should happen instantly (if not already in a presentation period) to fill the seat if removal means that the desired members are not met.

pub fn set_presentation_duration(
origin: OriginFor<T>,
count: T::BlockNumber
) -> DispatchResult

Set the presentation duration. If there is currently a vote being presented for, will invoke finalize_vote.

pub fn set_term_duration(
origin: OriginFor<T>,
count: T::BlockNumber
) -> DispatchResult

Set the presentation duration. If there is current a vote being presented for, will invoke finalize_vote.

impl<T: Config> Pallet<T>

pub fn presentation_duration() -> T::BlockNumber

How long to give each top candidate to present themselves after the vote ends.

impl<T: Config> Pallet<T>

pub fn term_duration() -> T::BlockNumber

How long each position is active for.

impl<T: Config> Pallet<T>

pub fn desired_seats() -> u32

Number of accounts that should constitute the collective.

impl<T: Config> Pallet<T>

pub fn members() -> Vec<(T::AccountId, T::BlockNumber)>

The current membership. When there’s a vote going on, this should still be used for executive matters. The block number (second element in the tuple) is the block that their position is active until (calculated by the sum of the block number when the member was elected and their term duration).

impl<T: Config> Pallet<T>

pub fn vote_index() -> VoteIndex

The total number of vote rounds that have happened or are in progress.

impl<T: Config> Pallet<T>

pub fn approvals_of<KArg>(k: KArg) -> Vec<u32> where
KArg: EncodeLike<(T::AccountId, u32)>,

TWOX-NOTE: SAFE as AccountId is a crypto hash and SetIndex is not attacker-controlled.

impl<T: Config> Pallet<T>

pub fn candidate_reg_info<KArg>(k: KArg) -> Option<(VoteIndex, u32)> where
KArg: EncodeLike<T::AccountId>,

The vote index and list slot that the candidate who was registered or None if they are not currently registered.

TWOX-NOTE: SAFE as AccountId is a crypto hash.

impl<T: Config> Pallet<T>

pub fn voter_info<KArg>(
k: KArg
) -> Option<VoterInfo<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>> where
KArg: EncodeLike<T::AccountId>,

Basic information about a voter.

TWOX-NOTE: SAFE as AccountId is a crypto hash.

impl<T: Config> Pallet<T>

pub fn voters<KArg>(k: KArg) -> Vec<Option<T::AccountId>> where
KArg: EncodeLike<u32>,

The present voter list (chunked and capped at VOTER_SET_SIZE).

TWOX-NOTE: OKAY ― SetIndex is not user-controlled data.

impl<T: Config> Pallet<T>

pub fn next_nonfull_voter_set() -> u32

the next free set to store a voter in. This will keep growing.

impl<T: Config> Pallet<T>

pub fn voter_count() -> u32

Current number of Voters.

impl<T: Config> Pallet<T>

pub fn candidates() -> Vec<T::AccountId>

The present candidate list.

impl<T: Config> Pallet<T>

pub fn candidate_count() -> u32

Current number of active candidates

impl<T: Config> Pallet<T>

pub fn next_finalize() -> Option<(T::BlockNumber, u32, Vec<T::AccountId>)>

The accounts holding the seats that will become free on the next tally.

impl<T: Config> Pallet<T>

pub fn leaderboard(
) -> Option<Vec<(<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, T::AccountId)>>

Get the leaderboard if we’re in the presentation phase. The first element is the weight of each entry; It may be the direct summed approval stakes, or a weighted version of it. Sorted from low to high.

impl<T: Config> Pallet<T>

pub fn presentation_active() -> bool

True if we’re currently in a presentation period.

pub fn is_a_candidate(who: &T::AccountId) -> bool

If who a candidate at the moment?

pub fn will_still_be_member_at(who: &T::AccountId, n: T::BlockNumber) -> bool

Iff the member who still has a seat at blocknumber n returns true.

pub fn next_vote_from(n: T::BlockNumber) -> T::BlockNumber

Determine the block that a vote can happen on which is no less than n.

pub fn next_tally() -> Option<T::BlockNumber>

The block number on which the tally for the next election will happen. None only if the desired seats of the set is zero.

pub fn bool_to_flag(x: Vec<bool>) -> Vec<u32>

Convert a vec of boolean approval flags to a vec of integers, as denoted by the type ApprovalFlag. see bool_to_flag_should_work test for examples.

pub fn flag_to_bool(chunk: Vec<u32>) -> Vec<bool>

Convert a vec of flags (u32) to boolean.