Crate pallet_elections

WARNING: NOT ACTIVELY MAINTAINED

This pallet is currently not maintained and should not be used in production until further notice.

Election pallet for stake-weighted membership selection of a collective.

The composition of a set of account IDs works according to one or more approval votes weighted by stake. There is a partial carry-over facility to give greater weight to those whose voting is serially unsuccessful.

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

VoterInfo
The activity status of a voter.

Enums

CellStatus
Used to demonstrate the status of a particular index in the global voter list.

Constants

APPROVAL_SET_SIZE
NUmber of approvals grouped in one chunk.

VOTER_SET_SIZE
Number of voters grouped in one chunk.

Type Definitions

VoteIndex
Index used to count voting rounds.